feat: 하드코딩 된 부분 삭제 및 테스트 환경 설정

+++ src/main/java/com/command/itdaserver/domain/auth/domain/UserSession.java
@@ -28,9 +28,9 @@ public class UserSession {
     private String role;
 
     @TimeToLive
-    private Long expiration; // TTL (초 단위) - 예: 1800 (30분)
+    private int expiration; // TTL (초 단위) - 예: 1800 (30분)
 
-    public static UserSession create(String userId, String email, String role, Long expiration) {
+    public static UserSession create(String userId, String email, String role, int expiration) {
         return UserSession.builder()
                 .sessionId(UUID.randomUUID().toString()) // 랜덤 세션 ID
                 .userId(userId)
