+++ src/main/java/com/command/itdaserver/domain/auth/service/LoginService.java
@@ -24,7 +24,7 @@ public class LoginService {
     private final SessionRepository sessionRepository;
 
     @Value("${session.expiration:1800}")
-    private Long sessionExpiration;
+    private int sessionExpiration;
 
     public String execute(LoginRequest request){
         log.info("로그인 시도 - userId: {}", request.userId());
